split FileUploadField into multiple components
