+++ demo/site/src/common/components/form/ChooseButton.module.scss
@@ -1,0 +1,28 @@
+.chooseButton {
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    gap: var(--spacing-s100);
+    width: 100%;
+    height: 48px;
+    padding: var(--spacing-s400);
+    border-radius: 4px;
+    border: 1px solid var(--cta-outlined-default-border);
+    background-color: var(--cta-outlined-default-bg, transparent);
+    color: var(--cta-outlined-default-text);
+    cursor: pointer;
+    box-sizing: border-box;
+    font-size: 16px;
+    font-weight: 700;
+    line-height: 110%;
+    transition:
+        background-color 0.2s ease-out,
+        color 0.2s ease-out,
+        border-color 0.2s ease-out;
+
+    &:hover {
+        background-color: var(--cta-outlined-hover-bg);
+        color: var(--cta-outlined-hover-text);
+        border-color: var(--cta-outlined-hover-border);
+    }
+}
