+++ demo/site/src/common/components/form/FileItem.module.scss
@@ -1,0 +1,50 @@
+.fileItem {
+    display: flex;
+    align-items: center;
+    gap: var(--spacing-s100);
+    height: 40px;
+    padding: var(--spacing-s200);
+    border-radius: 4px;
+    border: 1px solid var(--contact-form-border-border);
+    background-color: var(--contact-form-bg-bg, white);
+    color: var(--contact-form-text-input-text);
+    box-sizing: border-box;
+    font-size: 13px;
+    line-height: 20px;
+
+    &--error {
+        border-color: var(--contact-form-border-error);
+        color: var(--contact-form-text-error);
+    }
+}
+
+.fileIcon {
+    flex-shrink: 0;
+    color: var(--contact-form-icon-default);
+}
+
+.fileItem--error .fileIcon {
+    color: var(--contact-form-border-error);
+}
+
+.fileName {
+    flex: 1;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+}
+
+.removeButton {
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    flex-shrink: 0;
+    width: 24px;
+    height: 24px;
+    padding: 0;
+    border: 0;
+    border-radius: 4px;
+    background: none;
+    color: var(--contact-form-icon-default);
+    cursor: pointer;
+}
