+++ demo/site/src/common/components/form/FileUploadField.module.scss
@@ -16,35 +16,6 @@
     border: 0;
 }
 
-.chooseButton {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    gap: var(--spacing-s100);
-    width: 100%;
-    height: 48px;
-    padding: var(--spacing-s400);
-    border-radius: 4px;
-    border: 1px solid var(--cta-outlined-default-border);
-    background-color: var(--cta-outlined-default-bg, transparent);
-    color: var(--cta-outlined-default-text);
-    cursor: pointer;
-    box-sizing: border-box;
-    font-size: 16px;
-    font-weight: 700;
-    line-height: 110%;
-    transition:
-        background-color 0.2s ease-out,
-        color 0.2s ease-out,
-        border-color 0.2s ease-out;
-
-    &:hover {
-        background-color: var(--cta-outlined-hover-bg);
-        color: var(--cta-outlined-hover-text);
-        border-color: var(--cta-outlined-hover-border);
-    }
-}
-
 .fileList {
     list-style: none;
     margin: 0;
@@ -52,55 +23,4 @@
     display: flex;
     flex-direction: column;
     gap: var(--spacing-s100);
-}
-
-.fileItem {
-    display: flex;
-    align-items: center;
-    gap: var(--spacing-s100);
-    height: 40px;
-    padding: var(--spacing-s200);
-    border-radius: 4px;
-    border: 1px solid var(--contact-form-border-border);
-    background-color: var(--contact-form-bg-bg, white);
-    color: var(--contact-form-text-input-text);
-    box-sizing: border-box;
-    font-size: 13px;
-    line-height: 20px;
-
-    &--error {
-        border-color: var(--contact-form-border-error);
-        color: var(--contact-form-text-error);
-    }
-}
-
-.fileIcon {
-    flex-shrink: 0;
-    color: var(--contact-form-icon-default);
-}
-
-.fileItem--error .fileIcon {
-    color: var(--contact-form-border-error);
-}
-
-.fileName {
-    flex: 1;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-}
-
-.removeButton {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    flex-shrink: 0;
-    width: 24px;
-    height: 24px;
-    padding: 0;
-    border: 0;
-    border-radius: 4px;
-    background: none;
-    color: var(--contact-form-icon-default);
-    cursor: pointer;
 }
